Ajuste de rotas, inclusão de menus e criação da screen Graff
