+++ FrontEnd/App.js
@@ -1,4 +1,7 @@
+import 'react-native-gesture-handler';
+
 import Routes from "./src/Routes";
+
 
 export default function App() {
   return (
