--- FrontEnd/src/pages/InsertProperty/index.js
@@ -1,9 +1,0 @@
-import {Text, View} from 'react-native'
-
-export default function InsertPropertyScreen() {
-  return (
-    <View>
-        <Text> Cadastrar propriedade </Text>
-    </View>
-  );
-}
