+++ FrontEnd/src/pages/PerfilGrafiteiro/index.js
@@ -1,0 +1,9 @@
+import {Text, View} from 'react-native'
+
+export default function PerfilGrafiteiroScreen() {
+  return (
+    <View>
+        <Text> feed de muros disponíveis </Text>
+    </View>
+  );
+}
